extract password controller
